+++ src/main/java/me/andromedov/mythicskript/expressions/mythicitem/GetMythicTypeOfItem.java
@@ -1,7 +1,5 @@
 package me.andromedov.mythicskript.expressions.mythicitem;
 
-import javax.annotation.Nullable;
-
 import org.bukkit.event.Event;
 import org.bukkit.inventory.ItemStack;
 
@@ -34,7 +32,7 @@ public class GetMythicTypeOfItem extends SimpleExpression<String> {
     }
 
     @Override
-    public String toString(@Nullable Event e, boolean debug) {
+    public String toString(Event e, boolean debug) {
         return getClass().getSimpleName() + (e != null ? "@" + e.getEventName() : "");
     }